+++ index.js
@@ -7,23 +7,23 @@ module.exports = {
     'stylelint-order',
     'stylelint-prettier'
   ],
-  extends: [
-    'stylelint-config-standard',
-    'stylelint-prettier/recommended'
-  ],
+  extends: ['stylelint-config-standard', 'stylelint-prettier/recommended'],
   rules: {
-    'at-rule-no-unknown': [true, {
-      ignoreAtRules: ['function', 'if', 'else', 'each', 'include', 'mixin', 'return', 'for', 'use']
-    }],
-    'at-rule-blacklist': ['extend', 'import'],
+    'at-rule-blacklist': [
+      // Disable @extend in favour of @mixin
+      'extend',
+      // Disable @import in favour of @user
+      'import'
+    ],
+    // Colours should always reference a variable
     'color-no-hex': true,
-    'color-hex-length': 'long',
+    // There is always a better alternative to using !important
     'declaration-no-important': true,
-    'font-family-name-quotes': null,
-    'keyframes-name-pattern': pattern,
+    // Deep nesting creates overly specified rules
     'max-nesting-depth': 3,
-    'no-descending-specificity': null, // enable me soon
-    'number-max-precision': 3,
+    // Enable me soon
+    'no-descending-specificity': null,
+    // Make sure @ rules come at the beginning
     'order/order': [
       {
         type: 'at-rule',
@@ -36,6 +36,7 @@ module.exports = {
       },
       'declarations'
     ],
+    // Use a loose BEM naming convention by @define'ing a namespace
     'plugin/selector-bem-pattern': {
       componentName: pattern,
       componentSelectors: {
@@ -43,15 +44,26 @@ module.exports = {
         combined: '.*'
       }
     },
-    'property-no-vendor-prefix': true,
-    'selector-class-pattern': [ pattern, { resolveNestedSelectors: true } ],
+    // Use a loose naming convention for animation keyframe names
+    'keyframes-name-pattern': pattern,
+    // Use a loose BEM naming convention for class names
+    'selector-class-pattern': [pattern, { resolveNestedSelectors: true }],
+    // Use a loose naming convention for ID selectors
     'selector-id-pattern': pattern,
+    // Disallow vendor prefixes, these should be added by autoprefixer
+    'property-no-vendor-prefix': true,
+    // Try to avoid complex selectors
     'selector-max-compound-selectors': 5,
+    // No need to specify IDs, class names will suffice
     'selector-max-id': 0,
+    // Disallow styling global elements, use BEM instead
     'selector-max-type': 0,
+    // Disallow styling everything at once
     'selector-max-universal': 0,
-    'selector-no-qualifying-type': [true, { ignore: ['attribute'] }],
+    // No need to specify redundant values
     'shorthand-property-no-redundant-values': true,
+    // Enforce the use of a variable for these properties
+    // to provide consistancy.
     'sh-waqar/declaration-use-variable': [
       [
         '/color/',
@@ -73,8 +85,6 @@ module.exports = {
           ]
         }
       ]
-    ],
-    'unit-blacklist': ['rem', 'em'],
-    'value-keyword-case': 'lower'
+    ]
   }
 };